Move routes to a separate folder
